# packages/api/openapi.yaml
openapi: "3.0.3"
info:
  title: "iQbe API"
  version: 2.1.0
servers:
  - url: "http://localhost:9000/api/v2"
...
security:
  - jwt: []
paths:
  "/analysis/ranking/all/{period}": 
    get: 
      tags: 
        - analysis
      summary: ユーザの演習問題数のランキング
      description: 演習問題によるランク済みのユーザリストを取得する
      parameters:
        - name: period
          in: path
          required: true
          description: 日 / 週 / 月
          schema:
            type: string
          example: day 
      responses:
        "200": 
          description: 指定範囲のユーザ情報とランク
          content:
            application/json:
              schema: 
                type: array
                items: 
                  $ref: "#/components/schemas/Rank"
              example:
                - rank: 1
                  uid: WI0Nulr6F1dhLUi7A3zYRKs5bu9txaGgT0PqBc 
                  name: foo
                  count: 323
                  compare: 0
                - rank: 2
                  uid: 2kYfkh59sP8Qbm1drC7uZXOnBIGRSAoz5hTMN3
                  name: piyo
                  count: 234
                  compare: 1
                - rank: 3
                  uid: Kb5F7yQHOhpNb9tnLBx6GJj1IeiucMk0mDw27z
                  name: bar
                  count: 103
                  compare: -1
                - rank: 4
                  uid: BapXVEOuLlYb3jHCBpeSzx7cPRrgf1dNTZqE4
                  name: hoge
                  count: 234
                  compare: 1
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
  "/analysis/ranking/personal/{period}": 
    get: 
      tags: 
        - analysis
      summary: 個人の演習問題数のランク
      description: 演習問題によるランクを取得する
      parameters:
        - name: period
          in: path
          required: true
          description: 日 / 週 / 月
          schema:
            type: string
          example: day 
      responses:
        "200": 
          description: 指定範囲のランク
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Rank"
              example:
                rank: 1
                uid: WI0Nulr6F1dhLUi7A3zYRKs5bu9txaGgT0PqBc 
                name: foo
                count: 323
                compare: 0
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
  "/analysis/status/{date}/{period}": 
    get: 
      tags: 
        - analysis
      summary: 正答/誤答/スルー問題数を取得
      description: 指定する日時範囲の正答/誤答/スルー問題数を取得
      parameters:
        - name: date
          in: path
          required: true
          description: 起算する日にち
          schema:
            type: string
          example: 2023-12-31 
        - name: period
          in: path
          required: true
          description: 日 / 週 / 月
          schema:
            type: string
          example: day 
      responses:
        "200": 
          description: 正答/誤答/スルーの問題数
          content:
            application/json:
              schema:
                type: array
                items: 
                  type: object
                  properties:
                    start: 
                      type: string
                    end: 
                      type: string
                    right: 
                      type: number
                    wrong: 
                      type: number
                    through: 
                      type: number
                  required:
                    - start
                    - end
                    - right
                    - wrong
                    - through
              example:
                - start: 2023-12-31 00:00:00
                  end: 2023-12-31 23:59:59
                  right: 2
                  wrong: 1
                  through: 4
                - start: 2023-12-30 00:00:00
                  end: 2023-12-30 23:59:59
                  right: 10
                  wrong: 10
                  through: 40
                - start: 2023-12-29 00:00:00
                  end: 2023-12-29 23:59:59
                  right: 1
                  wrong: 3
                  through: 47
                - start: 2023-12-28 00:00:00
                  end: 2023-12-28 23:59:59
                  right: 17
                  wrong: 12
                  through: 37
                - start: 2023-12-27 00:00:00
                  end: 2023-12-27 23:59:59
                  right: 18
                  wrong: 14
                  through: 34
                - start: 2023-12-26 00:00:00
                  end: 2023-12-26 23:59:59
                  right: 13
                  wrong: 4
                  through: 12
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
...
components:
...
  responses:
    400BadRequest:
      description: パラメータの不備
      content:
        application/problem+json:
          schema: 
            $ref: "#/components/schemas/Error"
          examples:
            Invaild parameter: 
              value:
                title: INVAILD_PARAMS
                type: "about:blank"
                status: 400
                detail: "This request contains invaild parameter. Please correct to the right ones."
    401Unauthorized: 
      description: JWTトークンの不備（期限切れ，無効，存在しない）
      content:
        application/problem+json:
          schema: 
            $ref: "#/components/schemas/Error"
          examples: 
            Expired token:
              value:
                title: EXPIRED_TOKEN
                type: "about:blank"
                status: 401
                detail: "This authorization token have expired. Please refetch token."
            Invalid token:
              value:
                code: INVAILD_TOKEN
                type: "about:blank"
                status: 401
                detail: "This authorization token is invalid."
            No token:
              value:
                code: NoToken
                type: "about:blank"
                status: 401
                message: "Authorization token was not found. Please set authorization header."
    403Forbidden: 
      description: アクセス権限がない
      content:
        application/problem+json:
          schema: 
            $ref: "#/components/schemas/Error"
          examples:
            Access denied:
              value: 
                title: ACCESS_DENIED
                type: "about:blank"
                status: 403
                detail: "Access denied. Insufficient rights to access the resource."
...
  schemas:
...
    Error: 
      type: object
      properties: 
        title:
          type: string
          description: エラーを簡潔に表現したコード
        type: 
          type: string
          description: エラーの詳細ドキュメントを指すURL
        status:
          type: number
          description: HTTPステータスコード
        detail: 
          type: string
          description: 開発者向けの詳細文
      required:
        - title
        - type
        - status
        - detail
...
    Rank: 
      type: object
      properties:
        rank:
          type: number
        uid: 
          type: string
        username:
          type: string
        nickname:
          type: string
        count: 
          type: number
        compare: 
          type: number
      required:
        - rank
        - uid
        - username
        - count 
        - compare 
...
  securitySchemes:
    jwt:
      type: http
      description: Credentials or access token for API
      scheme: bearer
      bearerFormat: JWT